fix(backend): corrigir compilacao BigDecimal.valueOf(BigDecimal) em ApplianceAggregationService
getAverageDailyUseHours() ja retorna BigDecimal, entao
BigDecimal.valueOf(pa.getAppliance().getAverageDailyUseHours())
era invalido (BigDecimal.valueOf aceita long ou double,
nao BigDecimal). Usar o retorno diretamente.

Corrige falha nos workflows:
- lint-backend / static-analysis (spotless:check + compile)
- deploy-backend-dev / build-and-push (Docker build)

+++ backend/src/main/java/br/com/group18/energiai/application/services/ApplianceAggregationService.java
@@ -38,7 +38,7 @@ public class ApplianceAggregationService {
                     double monthlyKwh = pa.getMonthlyConsumptionKwh() != null
                             ? pa.getMonthlyConsumptionKwh().doubleValue()
                             : pa.getTotalPowerWatts()
-                                    .multiply(BigDecimal.valueOf(pa.getAppliance().getAverageDailyUseHours()))
+                                    .multiply(pa.getAppliance().getAverageDailyUseHours())
                                     .divide(BigDecimal.valueOf(1000), RoundingMode.HALF_UP)
                                     .multiply(BigDecimal.valueOf(30))
                                     .doubleValue();
